+++ types/src/commonMain/kotlin/kotools/types/collection/Error.kt
@@ -1,4 +1,4 @@
-package kotools.types
+package kotools.types.collection
 
 internal object EmptyCollectionError :
     IllegalArgumentException("Given collection shouldn't be empty.")