+++ types/src/commonMain/kotlin/kotools/types/collection/NotEmptySet.kt
@@ -1,10 +1,12 @@
-package kotools.types
+package kotools.types.collection
 
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.builtins.SetSerializer
 import kotools.shared.Project.Types
 import kotools.shared.SinceKotools
+import kotools.types.Serializer
+import kotools.types.toSuccessfulResult
 
 /**
  * Representation of sets that contain at least one element.
@@ -12,7 +14,7 @@ import kotools.shared.SinceKotools
  * @param E The type of elements contained in this set.
  */
 @Serializable(NotEmptySetSerializer::class)
-@SinceKotools(Types, "3.2")
+@SinceKotools(Types, "4.0")
 public class NotEmptySet<out E>
 private constructor(private val elements: Set<E>) : Set<E> by elements {
     internal companion object {
@@ -26,18 +28,11 @@ private constructor(private val elements: Set<E>) : Set<E> by elements {
     override fun toString(): String = "$elements"
 }
 
-internal class NotEmptySetSerializer<E>(elementSerializer: KSerializer<E>) :
-    Serializer<NotEmptySet<E>, Set<E>>(
-        delegate = SetSerializer(elementSerializer),
-        toDelegatedType = { it },
-        toType = Set<E>::toNotEmptySet
-    )
-
 /**
  * Creates a [NotEmptySet] starting with a [head] and containing all the
  * elements of the optional [tail].
  */
-@SinceKotools(Types, "3.2")
+@SinceKotools(Types, "4.0")
 public fun <E> notEmptySetOf(head: E, vararg tail: E): NotEmptySet<E> {
     val result: List<E> = listOf(head) + tail
     return result.toNotEmptySet()
@@ -48,6 +43,13 @@ public fun <E> notEmptySetOf(head: E, vararg tail: E): NotEmptySet<E> {
  * Returns a [NotEmptySet] containing all the elements of this collection, or
  * an [IllegalArgumentException] if this collection is empty.
  */
-@SinceKotools(Types, "3.2")
+@SinceKotools(Types, "4.0")
 public fun <E> Collection<E>.toNotEmptySet(): Result<NotEmptySet<E>> =
     NotEmptySet of this
+
+internal class NotEmptySetSerializer<E>(elementSerializer: KSerializer<E>) :
+    Serializer<NotEmptySet<E>, Set<E>>(
+        delegate = SetSerializer(elementSerializer),
+        toDelegatedType = { it },
+        toType = Set<E>::toNotEmptySet
+    )