+++ types/src/commonTest/kotlin/kotools/types/Assertions.kt
@@ -1,0 +1,8 @@
+package kotools.types
+
+import kotools.assert.assertNotNull
+import kotools.assert.assertTrue
+
+internal fun Throwable.assertHasAMessage(): Unit = message.assertNotNull()
+    .isNotBlank()
+    .assertTrue()